requirements.txt
# Core scientific + data
networkx
pandas
datasets

# Multi-agent frameworks
langgraph
crewai
gdm-concordia

# LLM utilities
python-dotenv

# Remote LLMs
openai
anthropic
google-generativeai   # Gemini

# Local LLMs
ollama
langchain-ollama

# Optional helpers
tqdm
scipy
langchain-openai
